Overhaul deck backup
Archidekt made major changes to their API, so the method we were using
to retrieve decks with their folder info doesn't work anymore.
With this update, we can backup all decks again, but they will only be
in a folder if the root of that folder tree is specified in the config.

Also added new deck formats to the format key, and added a new backup
format, which is a more minimal info version of JSON. The old version
of JSON is now called "rawJson" instead.